Issue 822
Cleaning up build warnings.

git-svn-id: http://tungsten-replicator.googlecode.com/svn/trunk@2044 9f5e3340-5c9a-25cb-7c1c-c133a2e9e928

+++ replicator/src/java/com/continuent/tungsten/replicator/consistency/DataScanCtrl.java
@@ -891,8 +891,8 @@ public class DataScanCtrl
         Table tableSlave = slaveDb.findTable(schema, table);
         ConsistencyCheck ccSlave = ConsistencyCheckFactory
                 .createConsistencyCheck(cc.getCheckId(), tableSlave,
-                        (int) cc.getRowOffset(), (int) cc.getRowLimit(),
-                        getMethod(), false, false);
+                        cc.getRowOffset(), cc.getRowLimit(), getMethod(),
+                        false, false);
 
         // Retrieve check results from the master.
         String query = String.format("SELECT %s,%s FROM %s.%s WHERE %s = %d",
